+++ pkg/collect/collector_resource.go
@@ -26,6 +26,9 @@ import (
 	"strconv"
 	goStrings "strings"
 
+	"golang.org/x/sys/unix"
+
+	shared "github.com/arangodb/kube-arangodb/pkg/apis/shared"
 	"github.com/arangodb/kube-arangodb/pkg/util"
 	utilConstants "github.com/arangodb/kube-arangodb/pkg/util/constants"
 	"github.com/arangodb/kube-arangodb/pkg/util/errors"
@@ -48,11 +51,18 @@ const (
 	metricCPULimits      = "cpu_limits"
 	metricMemoryRequests = "memory_requests"
 	metricMemoryLimits   = "memory_limits"
+
+	// metricDataStorage is the total size of the arangod data volume (/data) in GB.
+	metricDataStorage = "data_storage"
+
+	// metricDataStorageAvailable is the free space on the arangod data volume (/data) in GB.
+	metricDataStorageAvailable = "data_storage_available"
 )
 
 func init() {
 	GetCollector().Register(resourceCollector{})
 	GetCollector().Register(resourceRequestsLimitsCollector{})
+	GetCollector().Register(dataStorageCollector{})
 }
 
 // resourceRequestsLimitsCollector pushes the container's CPU/memory requests and limits (from the
@@ -83,6 +93,47 @@ func (resourceRequestsLimitsCollector) CollectEvents(out util.Pusher[Metric]) er
 	}
 
 	return nil
+}
+
+// dataStorageCollector pushes the total and available size of the arangod data volume (/data) as event
+// body metrics (in GB). It is best-effort: when the data directory is not present (e.g. the collector
+// runs in a container without the data volume mounted, such as the gateway) the metrics are skipped.
+type dataStorageCollector struct{}
+
+func (dataStorageCollector) CollectEvents(out util.Pusher[Metric]) error {
+	total, available, exists, err := dataStorage(shared.ArangodVolumeMountDir)
+	if err != nil {
+		return err
+	}
+	if !exists {
+		// No data volume mounted (e.g. gateway); nothing to report.
+		return nil
+	}
+
+	out.Push(Metric{K: metricDataStorage, V: float32(total) / bytesPerGB})
+	out.Push(Metric{K: metricDataStorageAvailable, V: float32(available) / bytesPerGB})
+
+	return nil
+}
+
+// dataStorage returns the total and available size (in bytes) of the filesystem backing path. exists
+// is false when the path does not exist (the metrics are then skipped).
+func dataStorage(path string) (total, available uint64, exists bool, err error) {
+	if _, err := os.Stat(path); err != nil {
+		if os.IsNotExist(err) {
+			return 0, 0, false, nil
+		}
+		return 0, 0, false, errors.Wrapf(err, "unable to stat %s", path)
+	}
+
+	var stat unix.Statfs_t
+	if err := unix.Statfs(path, &stat); err != nil {
+		return 0, 0, false, errors.Wrapf(err, "unable to statfs %s", path)
+	}
+
+	blockSize := uint64(stat.Bsize)
+
+	return stat.Blocks * blockSize, stat.Bavail * blockSize, true, nil
 }
 
 // resourceCollector pushes the available CPU and memory as event body metrics.